+++ src/app/dto/entity/Author.java
@@ -1,6 +1,5 @@
-package app.dto;
+package app.dto.entity;
 
-import app.enums.CategoryStateEnum;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,12 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class Category {
+public class Author {
 	private int sequence;
 	private String name;
-	private int level;
-	private CategoryStateEnum state;
-	
-	// FK
-	private int parentCategory;
+
 }
