+++ src/app/dto/entity/Category.java
@@ -1,5 +1,6 @@
-package app.dto;
+package app.dto.entity;
 
+import app.enums.CategoryStateEnum;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -9,8 +10,12 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class Publisher {
+public class Category {
 	private int sequence;
 	private String name;
-
+	private int level;
+	private CategoryStateEnum state;
+	
+	// FK
+	private int parentCategory;
 }
