+++ src/app/dto/entity/OrderDetail.java
@@ -1,4 +1,4 @@
-package app.dto;
+package app.dto.entity;
 
 import app.enums.OrderStateEnum;
 import lombok.AllArgsConstructor;
