+++ src/app/dto/entity/Publisher.java
@@ -1,4 +1,4 @@
-package app.dto;
+package app.dto.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class Cust {
-	private String id;
-	private String pwd;
+public class Publisher {
+	private int sequence;
 	private String name;
+
 }
