+++ test/cust/CustSelect.java
@@ -4,7 +4,7 @@ package cust;
 import org.junit.Test;
 
 import app.cust.CustServiceImpl;
-import app.dto.Cust;
+import app.dto.entity.Cust;
 
 public class CustSelect {
 	@Test
